tests/wpt/web-platform-tests/css/css-sizing/aspect-ratio/parsing/contain-intrinsic-size-computed.html

<!DOCTYPE html>
<title>CSS aspect-ratio Test: aspect-ratio with computed values</title>
<link rel="author" title="Google LLC" href="https://www.google.com/">
<link rel="help" href="https://drafts.csswg.org/css-sizing-4/#aspect-ratio">
<meta name="assert" content="aspect-ratio supports the full grammar '[auto | <ratio>]'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>

<div id=target></div>
<div id=scratch></div>
<script>

test_computed_value("aspect-ratio", "auto");
test_computed_value("aspect-ratio", "1");
test_computed_value("aspect-ratio", "1.3");
test_computed_value("aspect-ratio", "1 / 1", "1");
test_computed_value("aspect-ratio", "16 / 9");
test_computed_value("aspect-ratio", "16.3 / 9.5");
test_computed_value("aspect-ratio", "16/9", "16 / 9");
test_computed_value("aspect-ratio", "0 / 9");
test_computed_value("aspect-ratio", "16 / 0");
test_computed_value("aspect-ratio", "auto 1");
test_computed_value("aspect-ratio", "auto 1 / 1", "auto 1");
test_computed_value("aspect-ratio", "0 auto", "auto 0");
// The computed value of a <ratio> is the pair of numbers provided, unless
// both numbers are zero, in which case the computed value is the pair (1, 0)
// (same as 1 / 0).
// https://github.com/w3c/csswg-drafts/issues/5084
test_computed_value("aspect-ratio", "0 / 0", "1 / 0");
</script>
