chromium/third_party/blink/web_tests/external/wpt/css/css-sizing/aspect-ratio/parsing/contain-intrinsic-size-computed-expected.txt

This is a testharness.js-based test.
PASS Property aspect-ratio value 'auto'
FAIL Property aspect-ratio value '1' assert_true: '1' is a supported value for aspect-ratio. expected true got false
FAIL Property aspect-ratio value '1.3' assert_true: '1.3' is a supported value for aspect-ratio. expected true got false
FAIL Property aspect-ratio value '1 / 1' assert_equals: expected "1" but got "1 / 1"
PASS Property aspect-ratio value '16 / 9'
FAIL Property aspect-ratio value '16.3 / 9.5' assert_true: '16.3 / 9.5' is a supported value for aspect-ratio. expected true got false
PASS Property aspect-ratio value '16/9'
FAIL Property aspect-ratio value '0 / 9' assert_true: '0 / 9' is a supported value for aspect-ratio. expected true got false
FAIL Property aspect-ratio value '16 / 0' assert_true: '16 / 0' is a supported value for aspect-ratio. expected true got false
FAIL Property aspect-ratio value 'auto 1' assert_true: 'auto 1' is a supported value for aspect-ratio. expected true got false
FAIL Property aspect-ratio value 'auto 1 / 1' assert_true: 'auto 1 / 1' is a supported value for aspect-ratio. expected true got false
FAIL Property aspect-ratio value '0 auto' assert_true: '0 auto' is a supported value for aspect-ratio. expected true got false
FAIL Property aspect-ratio value '0 / 0' assert_true: '0 / 0' is a supported value for aspect-ratio. expected true got false
Harness: the test ran to completion.
